Feature/fetch real data for wallet hardcoded wallet address (#39)
* Rename a test and make another test less flaky by waiting longer for UI to update

* Stubbed out NftContentFetcher - specifically the constructor

* Rename a test and make another test less flaky by waiting longer for UI to update

* Ripped out Nethereum. Prepare dependencies for package exporting

* Wrap editor scripts in #if UNITY_EDITOR preprocessor defines so that they don't interfere with the build process

* Remove unneeded code

* Removed unneeded file

* Make UI in sample scene scale based on the screen size

* ContentFetcher implementation - including implementations for NftContentFetcher and TokenContentFetcher, which are essentially wrappers for ContentFetcher

* Fetch real fungible tokens.

* Added some logging to requests. Added error handling for HTTP requests sent by Indexer. Fixed error with content fetching by removing an unsupported Chain

* Fix tests - were checking for mock fetchers in the wrong place

* Fetch real nfts prototype

* Move SpriteFetcher.cs

* Finish fetching tokens for a chain first, then start processing collections asynchronously

* Retry failed requests to indexer due to rate limits. Fix some bugs

* Cleaner logging

* Move to next chain when receiving errors fetching from indexer - handles case when the indexer is down for a given chain. This logs a warning and points to the indexer status page as a first thing to check

* More informative logging

* Fix some bugs with fetching nfts. Uncaught exceptions were causing the async messages to abort silently

* Feed up fetch speed by adding nfts and tokens (fetching their images as well) to the lists in parallel

* Conducted trials with different number of tokens fetched at once to figure out an ideal starting place for devs to work with

* Fix broken UI tests

* Fixed bug with MockTransactionDetailsFetcher that was causing theUnity to freeze (infinite loop) when clicking on any of the wallet elements and trying to open their info page

* Destroy grid children on awake as opposed to everytime we open the walletpage

* Refresh ContentFetcher when re-opening the WalletPage

* Fixed bug where MockTransactionDetailsFetcher wasn't fetching transactions correctly

* Use my thread-safe implementation of DelayTask as opposed to the default Task.Delay

* Add basic end to end test for token and nft fetching and wallet page population. Fix memory leaks associated with web requests

* Don't swallow error when converting JSON objects

* Clean up ContentFetcher implementation

* Don't display NFTs that have the default sprite as their image in the wallet - i.e. those where we couldn't fetch the image

* Add icons in preparation for merge

+++ Assets/SequenceExamples/Scripts/Model/NftElement.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Numerics;
 using UnityEngine;
 
 namespace Sequence.Demo
@@ -8,12 +9,12 @@ namespace Sequence.Demo
         public CollectionInfo Collection;
         public Sprite TokenIconSprite;
         public string TokenName;
-        public uint TokenNumber;
+        public BigInteger TokenNumber;
         public uint Balance;
         public float EthValue;
         public ICurrencyConverter CurrencyConverter;
 
-        public NftElement(CollectionInfo collection, Sprite tokenIconSprite, string tokenName, uint tokenNumber, uint balance, float ethValue, ICurrencyConverter currencyConverter)
+        public NftElement(CollectionInfo collection, Sprite tokenIconSprite, string tokenName, BigInteger tokenNumber, uint balance, float ethValue, ICurrencyConverter currencyConverter)
         {
             Collection = collection;
             TokenIconSprite = tokenIconSprite;
@@ -25,7 +26,7 @@ namespace Sequence.Demo
         }
 
         public NftElement(Address contractAddress, Sprite tokenIconSprite, string tokenName,
-            Sprite collectionIconSprite, string collectionName, uint tokenNumber, Chain network, uint balance,
+            Sprite collectionIconSprite, string collectionName, BigInteger tokenNumber, Chain network, uint balance,
             float ethValue, ICurrencyConverter currencyConverter)
         {
             CollectionInfo collection = CollectionInfo.GetCollectionInfo(contractAddress, collectionIconSprite, collectionName, network);
